# modules/rest-api/src/rest-api-swagger.yaml
swagger: '2.0'
info:
  version: "6.3"
  title: CUBA Platform REST API
  description: |

    ## General Overview

    CUBA Platform REST API provides the following functionality:

    * Entities CRUD operations
    * Execution of predefined JPQL queries
    * Execution of services methods
    * Getting metadata (entities, views, enumerations, datatypes)
    * Getting current user permissions (access to entities, attributes, specific permissions)
    * Getting current user information (name, language, time zone, etc.)
    * Working with files

    The REST API is available at the URL `/rest/v2` both in the web and the portal modules, e.g. `http://localhost:8080/app/rest/v2/entities/sales$Order`

    ## Authentication

    ### Authentication using the OAuth2 protocol

    CUBA Platform REST API uses an OAuth2 resource owner password credentials grant to protect the endpoints (https://tools.ietf.org/html/rfc6749#page-37).

    Before accessing a protected resource, a user must get an authentication token. The token can be obtained at the `/oauth/token` URL.

    The `/oauth/token` endpoint is protected with the basic authentication. The request must contain an `Authorization` header for basic
    authentication with REST API client credentials. Please note that client credentials are not the same as platform user credentials.
    By client we mean an application that uses the REST API. The default client credentials are defined by the following
...

...

...

    ```
...
    ```

...

...

...

  license:
...

...
basePath: /rest/v2
produces:
...
consumes:
...
parameters:
  entityNameParam:
    name: entityName
    in: path
    type: string
    description: Entity name
    required: true
    default: ""
  entityIdParam:
    name: entityId
    in: path
    type: string
    description: Entity identifier
    required: true
    default: ""
  queryNameParam:
    name: queryName
    in: path
    type: string
    description: Query name
    required: true
    default: ""
...
  entityJsonParam:
    name: entityJson
    description: JSON object with the entity
    in: body
    schema:
      "$ref": "#/definitions/entity"
  entityViewParam:
    name: view
    in: query
    description: Name of the view which is used for loading the entity
    type: string
    default: ""
  returnNullsParam:
    name: returnNulls
    in: query
    description: Specifies whether null fields will be written to the result JSON
    type: boolean
  dynamicAttributesParam:
    name: dynamicAttributes
    in: query
    description: Specifies whether entity dynamic attributes should be returned
    type: boolean

################################################################################
#                                   Definitions                                #
################################################################################
definitions:
...
  error:
    type: object
    properties:
      error:
        type: string
        description: Error message
      details:
        type: string
        description: Detailed error description
...
  entity:
    type: object
    properties:
      id:
        type: string
        example: 4d8ef1d4-90b6-c075-8bfd-79181ff889ef
      otherEntityFields:
        type: string
        example: There must be a set of entity fields instead of this one here
  entityFull:
    type: object
    properties:
      id:
        type: string
        example: 4d8ef1d4-90b6-c075-8bfd-79181ff889ef
      _entityName:
        type: string
        description: Entity name
        example: sec$User
      _instanceName:
        type: string
        description: Entity instance name. This value must be used for displaying the entity in list.
        example: Smith [smith]
      otherEntityFields:
        type: string
        description: There must be a set of entity fields instead of this one here.
        example: There must be a set of entity fields instead of this one here
    required: [id, _entityName]
  predefinedQuery:
    type: object
    properties:
      name:
        type: string
        description: Query name
        example: usersByLastName
      jpql:
        type: string
        description: JPQL query
        example: select u from sec$User u where u.lastName = :lastName
      entityName:
        type: string
        description: Entity name
        example: sec$User
      viewName:
        type: string
        description: View name
        example: user.edit
      params:
        type: object
        description: Collection of query parameters
        properties:
          name:
            type: string
            description: Parameter name
            example: lastName
          type:
            type: string
            description: Parameter type. The fully qualified java class name or primitive type name is returned here.
            example: java.lang.String
...
  entityMetadata:
    type: object
    properties:
      entityName:
        type: string
        example: ref$Driver
      properties:
        type: array
        items:
          type: object
          properties:
            name:
              type: string
              example: name
            attributeType:
              type: string
              enum:
                - DATATYPE
                - ENUM
                - ASSOCIATION
                - COMPOSITION
            type:
              type: string
              example: string
              description: |
                For a simple datatype there will be a datatype name here (string, int, decimal, boolean, date, dateTime, etc.)
                You can find all simple datatype names in the source code of the Datatypes class in the systemDatatypeNames property.
                If the property is the link to the entity then an entity name will be placed to this field, e.g. ref$Colour
            cardinality:
              type: string
              enum:
                - NONE
                - ONE_TO_ONE
                - MANY_TO_ONE
                - ONE_TO_MANY
                - MANY_TO_MANY
            mandatory:
              type: boolean
            readOnly:
              type: boolean
            transient:
              type: boolean
...
responses:
  metaClassNotFoundResponse:
    description: Not found. MetaClass for the entity with the given name not found
    schema:
      $ref: '#/definitions/error'


################################################################################
#                                   Tags                                       #
################################################################################
...
securityDefinitions:
  oauth2:
    type: oauth2
    flow: password
    tokenUrl: http://localhost:8080/app/dispatch/api/oauth/token
    scopes:
      rest-api: use REST API
  tokenEndpointBasic:
    type: basic
security:
  - oauth2:
    - rest-api

################################################################################
#                                   Paths                                      #
################################################################################
paths:
...
  /entities/{entityName}:
    parameters:
      - $ref: '#/parameters/entityNameParam'
    get:
      tags:
        - Entities
      summary: Get a list of entities
      description: |
        Gets a list of entities
      parameters:
        - $ref: '#/parameters/entityViewParam'
        - name: limit
          in: query
          description: Number of extracted entities
          type: string
          default: ""
        - name: offset
          in: query
          description: Position of the first result to retrieve
          type: string
          default: ""
        - name: sort
          in: query
          description: |
            Name of the field to be sorted by. If the name is preceeding by the '+' character, then
            the sort order is ascending, if by the '-' character then descending. If there is no special
            character before the property name, then ascending sort will be used.
          type: string
          default: ""
        - $ref: '#/parameters/returnNullsParam'
        - $ref: '#/parameters/dynamicAttributesParam'

      responses:
        200:
          description: Success. The list of entities is returned in the response body.
          schema:
            type: array
            items:
              "$ref": "#/definitions/entityFull"
        403:
          description: Forbidden. The user doesn't have permissions to read the entity
          schema:
            $ref: '#/definitions/error'
        404:
          $ref: '#/responses/metaClassNotFoundResponse'
    post:
      tags:
        - Entities
      summary: Create new entity
      description: |
        Creates new entity. The method expects a JSON with entity object in the request body. The entity object
        may contain references to other entities. These references are processed according to the following rules:

        * If the entity property is annotated with @Composition in the entity java class, then it will be saved with the
        main entity
        * Otherwise a referenced entity with the given id will be searched. If it is found then the saved entity will
        have a reference to it. Otherwise, a response with code 400 will be returned.
      parameters:
        - name: entityJson
          description: |
            A JSON object with the entity
          in: body
          schema:
            "$ref": "#/definitions/entity"
      responses:
        201:
          description: Entity created. The created entity is returned in the response body.
          schema:
            $ref: '#/definitions/entityFull'
          headers:
            location:
              type: string
              description:
                A URL the created object can be accessed with
        400:
          description: Bad request. For example, the entity may have a reference to the non-existing entity.
          schema:
            $ref: '#/definitions/error'

        403:
          description: Forbidden. The user doesn't have permissions to create the entity
          schema:
            $ref: '#/definitions/error'
        404:
          $ref: '#/responses/metaClassNotFoundResponse'
  /entities/{entityName}/{entityId}:
    parameters:
      - $ref: '#/parameters/entityNameParam'
      - $ref: '#/parameters/entityIdParam'
    get:
      tags:
        - Entities
      summary: Get a single entity by id
      description: |
        Gets a single entity by identifier
      parameters:
        - $ref: '#/parameters/entityViewParam'
        - $ref: '#/parameters/returnNullsParam'
        - $ref: '#/parameters/dynamicAttributesParam'
      responses:
        200:
          description: Success. The entitiy is returned in the response body.
          schema:
            $ref: '#/definitions/entityFull'
        403:
          description: Forbidden. The user doesn't have permissions to read the entity
          schema:
            $ref: '#/definitions/error'
        404:
          description: |
            MetaClass not found or entity with the five identifier not found.
          schema:
            $ref: '#/definitions/error'
    put:
      tags:
        - Entities
      summary: Update the entity
      description: |
        Updates the entity. Only fields that are passed in the JSON object (the request body) are updated.
      parameters:
        - $ref: '#/parameters/entityJsonParam'
      responses:
        200:
          description: Success. The updated entity is returned in the response body.
          schema:
            $ref: '#/definitions/entityFull'
        403:
          description: Forbidden. The user doesn't have permissions to update the entity
          schema:
            $ref: '#/definitions/error'
        404:
          description: |
            MetaClass not found or entity with the given identifier not found.
          schema:
            $ref: '#/definitions/error'
    delete:
      tags:
        - Entities
      summary: Delete the entity
      description: Deletes the entity
      responses:
        200:
          description: Success. Entity was deleted.
        403:
          description: Forbidden. The user doesn't have permissions to delete the entity
          schema:
            $ref: '#/definitions/error'
        404:
          description: |
            MetaClass not found or entity with the given identifier not found.
          schema:
            $ref: '#/definitions/error'

############################### Queries ########################################
  /queries/{entityName}:
    parameters:
      - $ref: '#/parameters/entityNameParam'
    get:
      tags:
        - Queries
      summary: Get a list of queries
      description: |
        Gets a list of predefined queries for the entity
      responses:
        200:
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/predefinedQuery'
        403:
          description: Forbidden. The user doesn't have permissions to read the entity.
          schema:
            $ref: '#/definitions/error'
        404:
          description: |
            MetaClass not found.
  /queries/{entityName}/{queryName}:
    parameters:
      - $ref: '#/parameters/entityNameParam'
      - $ref: '#/parameters/queryNameParam'
      - name: limit
        in: query
        description: Number of extracted entities
        type: string
        default: ""
      - name: offset
        in: query
        description: Position of the first result to retrieve
        type: string
        default: ""
      - name: view
        in: query
        description: |
          Name of the view which is used for loading the entity. Specify this parameter if you want to extract entities with the view
          other than it is defined in the REST queries configuration file.
        type: string
      - $ref: '#/parameters/returnNullsParam'
      - $ref: '#/parameters/dynamicAttributesParam'
    get:
      tags:
        - Queries
      summary: Execute a query
      description: |
        Executes a predefined query. Query parameters values must be defined in the request parameters with the same name.
      responses:
        200:
          description: Success. A list of entitie is returned in the response body.
          schema:
            type: array
            items:
              $ref: '#/definitions/entityFull'
        403:
          description: Forbidden. The user doesn't have permissions to read the entity.
          schema:
            $ref: '#/definitions/error'
        404:
          description: |
            MetaClass not found or query with the given name not found
          schema:
            $ref: '#/definitions/error'
  /queries/{entityName}/{queryName}/count:
    parameters:
      - $ref: '#/parameters/entityNameParam'
      - $ref: '#/parameters/queryNameParam'
    get:
      tags:
        - Queries
      summary: Return a number of entities in query result
      description: |
        Returns a number of entities that matches the query. You can use the `all` keyword for the `queryNameParam`
        to get the number of all available entities.
      responses:
        200:
          description: Success
          schema:
            title: Count
            type: integer
        403:
          description: Forbidden. The user doesn't have permissions to read the entity.
          schema:
            $ref: '#/definitions/error'
        404:
          description: |
            MetaClass not found or query with the given name not found
          schema:
            $ref: '#/definitions/error'

############################### Services #######################################
...
  /metadata/entities/{entityName}:
    get:
      tags:
        - Metadata
      summary: Get entity metadata
      description: Gets entity metadata
      parameters:
        - $ref: '#/parameters/entityNameParam'
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/entityMetadata'
        404:
          description: |
            MetaClass not found or query with the given name not found
          schema:
            $ref: '#/definitions/error'
...